Build hygiene: fix the GCC-14 clean build, add check-standalone (W2.5)
Two things, both about build hygiene rather than the doomed src/ types (the
rebuilt core and shell already use <stdint.h>/<stdbool.h> only, carry no
BOOL/UCHAR/VOID/HANDLE, and build clean at -Wall -Wextra -Werror).

lib/rawhid/rawhid.c: fix the GCC-14 failure that broke `make clean && make`.
CM108Handle is a hid_device* (a real handle on Windows) but on Linux carries an
int fd stuffed into the pointer slot, then handed to read/write/close --
-Wint-conversion, a hard error by default in GCC 14. Recover the fd via
intptr_t at the three call sites (the standard fd-in-pointer idiom), which is
behaviour-preserving and satisfies the compiler. This is analysis/13 W2's "GCC
14 builds without -Wno-int-conversion" exit criterion; the inherited ardopcf now
builds clean from scratch under gcc 14.2.

Makefile + core/README.md: add `make check-standalone`, the mechanical form of
W2's "the core links with no ALSA / no sockets / no WebGui" criterion. It links
the whole core (codec + modem + templates + link) into a trivial program with
only -lm; a core object that ever grew a platform dependency fails the link
here. Documented as core rule 9, alongside check-pure and check-headers.

Verified: check-standalone passes, `make clean && make all` builds clean under
gcc 14, full sweep green (check-pure 19, check-headers, test-core 23, golden
424, golden-core), ardopb builds.

+++ lib/rawhid/rawhid.c
@@ -16,6 +16,7 @@
 
 #include <string.h>
 #include <stdio.h>
+#include <stdint.h>
 #include <errno.h>
 
 #include "hid/hidapi.h"
@@ -358,7 +359,10 @@ int HID_Read_Block()
 #ifdef WIN32
 	Len = rawhid_recv(0, Msg, 64, 100);
 #else
-	Len = read(CM108Handle, Msg, 64);
+	/* On Linux CM108Handle carries an int fd stuffed into the pointer slot
+	 * (Windows keeps a real hid_device*); recover it via intptr_t so GCC 14's
+	 * default -Wint-conversion error is satisfied. */
+	Len = read((int)(intptr_t)CM108Handle, Msg, 64);
 #endif
 
 	if (Len <= 0)
@@ -411,12 +415,12 @@ int HID_Write_Block()
 			return 0;
 		}
 #else
-		ret = write(CM108Handle, Msg, 64);
+		ret = write((int)(intptr_t)CM108Handle, Msg, 64);
 
 		if (ret != 64)
 		{
 			printf ("Write to %s failed, n=%d, errno=%d\n", HIDDevice, ret, errno);
-			close (CM108Handle);
+			close ((int)(intptr_t)CM108Handle);
 			CM108Handle = 0;
 			return 0;
 		}
